==> 2.4.Plot tables with number of matched articles and redirections. Save redirection bias groups.ipynb <==
from prettytable import PrettyTable

import json
def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)
    
words=load_simple_json('de/occupation_all.json')#{masculine:[[],[feminine]]}
words_revert={words[i][1]:i for i in words}#feminine:masculine
neutral=load_simple_json('de/neutral_cleaned.json')

no_page_m=load_simple_json('de/wiki/m_no_page.json')
no_page_f=load_simple_json('de/wiki/f_no_page.json')
no_page_n=load_simple_json('de/wiki/n_no_page.json')
page_m=load_simple_json('de/wiki/m_page.json')
page_f=load_simple_json('de/wiki/f_page.json')
page_n=load_simple_json('de/wiki/n_page.json')
m_page_levenshtein=load_simple_json('de/wiki/m_page_levenshtein.json')
#n_page_levenshtein=load_simple_json('de/wiki/n_page_levenshtein.json')


page_f_validated=load_simple_json('de/wiki/f_page_validated.json')
page_f_ambigious=load_simple_json('de/wiki/f_page_ambigious.json')
page_f_other=load_simple_json('de/wiki/f_page_other.json')  
    
page_m_validated=load_simple_json('de/wiki/m_page_validated.json') 
page_m_ambigious=load_simple_json('de/wiki/m_page_ambigious.json')
page_m_other=load_simple_json('de/wiki/m_page_other.json')

page_n_validated=load_simple_json('de/wiki/n_page_validated.json')
page_n_ambigious=load_simple_json('de/wiki/n_page_ambigious.json')
page_n_other=load_simple_json('de/wiki/n_page_other.json') 

m_links_to_feminine=load_simple_json('de/wiki/m_links_to_feminine.json')
m_links_to_feminine_lev=load_simple_json('de/wiki/m_links_to_feminine_lev.json')
m_links_to_smth=load_simple_json('de/wiki/m_links_to_smth.json')
f_links_to_mascuilne=load_simple_json('de/wiki/f_links_to_mascuilne.json')
f_links_to_smth=load_simple_json('de/wiki/f_links_to_smth.json')
redirection_n=load_simple_json('de/wiki/n_redirects.json')
n_links_to_masculine=load_simple_json('de/wiki/n_links_to_masculine.json')
n_links_to_smth=load_simple_json('de/wiki/n_links_to_smth.json')

f_links_to_mascuilne_lev=load_simple_json('de/wiki/f_links_to_mascuilne_lev.json')
f_links_to_mascuilne_with_link=load_simple_json('de/wiki/f_links_to_mascuilne_with_link.json')
f_redirection=load_simple_json('de/wiki/f_redirection.json')
m_redirection=load_simple_json('de/wiki/m_redirection.json')

t = PrettyTable(['-de-',"Occup.", 'Wiki pages', "Validated pages","Ambigious pages","Redirects",
                 "Redir. to oposit gender"])
t.add_row(['Masculine',len(words), len(page_m)+len(m_page_levenshtein), len(page_m_validated)+len(m_page_levenshtein),
           len(page_m_ambigious),
           len(m_redirection), len(m_links_to_feminine)])
t.add_row(['Feminine',len(words_revert), len(page_f), len(page_f_validated), len(page_f_ambigious),
           len(f_redirection),
           len(f_links_to_mascuilne)+len(f_links_to_mascuilne_lev)+len(f_links_to_mascuilne_with_link)])
t.add_row(['Neutral',len(neutral), len(page_n), len(page_n_validated), len(page_n_ambigious),
           len(redirection_n), "to M:"+str(len(n_links_to_masculine))+", to F:0"])
print t
t2 = PrettyTable(['-de-',"All", 'Wiki pages', "Redirects", "No page"])
t2.add_row(['Masculine',len(words), len(page_m)+len(m_page_levenshtein), len(m_redirection),
            len(no_page_m)])
t2.add_row(['Feminine',len(words_revert), len(page_f),
            len(f_redirection),
            len(no_page_f)])
t2.add_row(['Neutral',len(neutral), len(page_n),len(redirection_n),len(no_page_n)])
print t2
print
t3 = PrettyTable(['-de-','Wiki pages', "Validated pages","Ambigious pages","Not validated pages"])
t3.add_row(['Masculine', len(page_m)+len(m_page_levenshtein),
            len(page_m_validated)+len(m_page_levenshtein),
            len(page_m_ambigious),len(page_m_other)])
t3.add_row(['Feminine', len(page_f), len(page_f_validated), len(page_f_ambigious),len(page_f_other)])
t3.add_row(['Neutral', len(page_n), len(page_n_validated), len(page_n_ambigious),len(page_n_other)])

print t3
print
t4 = PrettyTable(['--de--',"Redirects","Redirects to oposit gender", "Other redirects"])
t4.add_row(['Masculine',len(m_redirection),
            len(m_links_to_feminine),len(m_links_to_smth)])
t4.add_row(['Feminine',len(f_redirection),
            len(f_links_to_mascuilne)+len(f_links_to_mascuilne_lev)+len(f_links_to_mascuilne_with_link),len(f_links_to_smth)])
t4.add_row(['Neutral',len(redirection_n),"to M:"+str(len(n_links_to_masculine))+", to F:0",len(n_links_to_smth)])
print t4



|    -de-   | Occup. | Wiki pages | Validated pages | Ambigious pages | Redirects | Redir. to oposit gender |
|:-:|:-:|:-:|:-:|:-:|:-:|:-:|
| Masculine |  4274  |    934     |       804       |       115       |    503    |            5            |
|  Feminine |  4274  |     20     |        13       |        5        |    310    |           214           |
|  Neutral  |  183   |     37     |        36       |        0        |     7     |      to M:3, to F:0     |


|           | All | Wiki pages | Redirects | No page |
|:-:|:-:|:-:|:-:|:-:|
| Masculine |  4274  |    934     |    503    |   2837  |
|  Feminine |  4274  |     20     |    310    |   3944  |
|  Neutral  |  183   |     37     |     7     |   139   |


|           | All wiki pages | Validated pages | Ambigious pages | Not validated pages |
|:-:|:-:|:-:|:-:|:-:|
| Masculine |    934     |       804       |       115       |          15         |
|  Feminine |     20     |        13        |        5        |          2          |
|  Neutral  |     37     |        36       |        0        |          1          |


|           | All wiki pages | Validated pages about profession | Other pages |
|:-:|:-:|:-:|:-:|:-:|
| Masculine |    934     |       815       |       119        |
|  Feminine |     20     |        18       |        2         |
|  Neutral  |     37     |        36       |        1         |


|           | All redirects | Redirects to opposite gender | Other redirects |
|:-:|:-:|:-:|:-:|
| Masculine |    503    |             5              |       498       |
|  Feminine |    310    |            214             |        96       |
|  Neutral  |     7     |       to M:3, to F:0       |        4        |



set(words_revert.keys())-set(no_page_f)-set(page_f)-set(f_links_to_mascuilne)-set(f_links_to_smth.keys())\
-set(f_links_to_mascuilne_with_link.keys())-set(f_links_to_mascuilne_lev.keys())

# do not start next one

from prettytable import PrettyTable
from tabulate import tabulate

import json
def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)


headers=["Occupation", 'Wiki pages', "Redirects","No wiki page or redir","All"]
t = PrettyTable(headers)


table=[]
for l in ["de","fr","ru"]:
    m_no_page=load_simple_json(l+'/wiki/m_no_page.json')
    m_redirection=load_simple_json(l+'/wiki/m_redirection.json')
    m_page=load_simple_json(l+'/wiki/m_page.json')
    words=len(m_no_page)+len(m_page)+ len(m_redirection)
    if l=="de":
        words_data=load_simple_json('de/occupation_all.json')
        words=len(words_data)
        m_page_levenshtein=load_simple_json(l+'/wiki/m_page_levenshtein.json')
        n_page=load_simple_json(l+'/wiki/n_page.json')
        n_no_page=load_simple_json(l+'/wiki/n_no_page.json')
        f_links_levenshtein=load_simple_json(l+'/wiki/f_links_to_mascuilne_lev.json')
        t.add_row(['{0}: Masculine'.format(l), len(m_page)+len(m_page_levenshtein),
                   len(m_redirection),len(m_no_page),words])
        f_no_page=load_simple_json(l+'/wiki/f_no_page.json')
        f_redirection=load_simple_json(l+'/wiki/f_redirection.json')
        f_page=load_simple_json(l+'/wiki/f_page.json')
        t.add_row(['{0}:  Feminine'.format(l),len(f_page),len(f_redirection),len(f_no_page), words])
        t.add_row(['{0}:  Neutral'.format(l),len(n_page),len(redirection_n),len(n_no_page), len(neutral)])
        table.append(['{0}: Masculine'.format(l), len(m_page)+len(m_page_levenshtein),
                   len(m_redirection),len(m_no_page),words])
        table.append(['{0}:  Feminine'.format(l), len(f_page),len(f_redirection),len(f_no_page),words])
    else:
        t.add_row(['{0}: Masculine'.format(l),len(m_page),len(m_redirection),len(m_no_page),words])
        table.append(['{0}: Masculine'.format(l),len(m_page),len(m_redirection),len(m_no_page),words])
print t

print tabulate(table, headers=headers, tablefmt="latex")#, numalign="left")



|Occupation  | Wiki pages | Redirects | No wiki page or redir | All  |
| :-|-|-|-|:-:|
| de: Masculine |    934     |    503    |          2837         | 4274 |
| de:  Feminine |     20     |    310    |          3944         | 4274 |
| de:   Neutral |     37     |      7    |          139          |  183 |
| fr: Masculine |     92     |     47    |          113          | 252  |
| ru: Masculine |    227     |    147    |          7296         | 7670 |


# Print all biases examples

f_b=[]
for i in words:
    if i in m_links_to_feminine:
        print "Female bias:", i ,"=redirects to=>", words[i][1]
        f_b.append(i)
    elif words[i][1] in page_f_ambigious:
        if (i not in page_m)&(i not in m_links_to_smth):
            print "Female bias:", i ,"=has only page=", words[i][1]
            f_b.append(i)
    elif words[i][1] in page_f_validated:
        if (i not in page_m)&(i not in m_links_to_smth):
            print "Female bias:", i ,"=has only page=", words[i][1]
            f_b.append(i)
print "\n\n"

n_b=[]
for i in words:
    if (i in page_m_validated)&(words[i][1] in page_f_validated):
        print "Neutral:", i ,"<= both pages exist=>", words[i][1]
        n_b.append(i)
    elif (i in page_m_ambigious)&(words[i][1] in page_f_validated):
        print "Neutral:", i ,"<= both pages exist=>", words[i][1]
        n_b.append(i)
    elif (i in m_links_to_smth)&(words[i][1] in page_f_validated):
        print "Neutral:", i ,"<=this redirects to smth ,both pages exist=>", words[i][1]
        n_b.append(i)
    elif words[i][1] in page_f_ambigious:
        if i in page_m:
            print "Neutral:", i ,"<= both pages exist & this is ambigious page=>", words[i][1]
            n_b.append(i)
        elif i in m_links_to_smth:
            print "Neutral:", i ,"<= this redirects to smth | this is ambigious page=>", words[i][1]
            n_b.append(i)
n_b2=[]
for i in page_n_validated:
    print "Neutral:", i 
    n_b2.append(i)
    
m_neutral_lev=load_simple_json('de/wiki/m_neutral_lev.json')    
    
for i in m_neutral_lev:
    if (i in no_page_m) and (i not in f_b):
        print "Neutral: field name of ",i, " =only exist= ", m_neutral_lev[i]
        n_b.append(i)
print len(f_b), len(n_b),len(n_b2)    


    
with open("de/wiki/f_bias.json","w")as f:
    json.dump(f_b, f, indent=4)
with open("de/wiki/n_bias_male_labels.json","w")as f:
    json.dump(n_b, f, indent=4)
with open("de/wiki/n_bias_n_labels.json","w")as f:
    json.dump(n_b2, f, indent=4)  

m_b=[]
for i in words:
    if words[i][1] in f_links_to_mascuilne:
        print "Male bias:", words[i][1] ,"=redirects to=>", i
        m_b.append(i)
    elif f_links_to_mascuilne_lev.has_key(words[i][1]):
        print "==Male bias:",words[i][1] ,"=redirects to=>", f_links_to_mascuilne_lev[words[i][1]].values()
        m_b.append(i)
    elif f_links_to_mascuilne_with_link.has_key(words[i][1]):
        print "=Male bias:",words[i][1] ,"=redirects to=>", f_links_to_mascuilne_with_link[words[i][1]][1]
        m_b.append(i)
    elif (i in page_m_validated) & (i not in n_b) & (i not in f_b):
        print "++Male bias: Only page ", i , " exist"
        m_b.append(i)
    elif i in n_links_to_masculine.values():
        #is already in male bias
        print "!!!!!!!!!!!!!!",i
    elif m_page_levenshtein.has_key(i):
        print "--Male bias: Only page ", i,m_page_levenshtein[i] , " exist"
        m_b.append(i)#27

print len(m_b)
with open("de/wiki/m_bias.json","w")as f:
    json.dump(m_b, f, indent=4) 

# Store all links in one file

import pandas as pd
import json
def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)
    
words=load_simple_json('de/occupation_all.json')#{masculine:[[],[feminine]]}
words_revert={words[i][1]:i for i in words}#feminine:masculine
neutral=load_simple_json('de/neutral_cleaned.json')

no_page_m=load_simple_json('de/wiki/m_no_page.json')
no_page_f=load_simple_json('de/wiki/f_no_page.json')
no_page_n=load_simple_json('de/wiki/n_no_page.json')
page_m=load_simple_json('de/wiki/m_page.json')
page_f=load_simple_json('de/wiki/f_page.json')
page_n=load_simple_json('de/wiki/n_page.json')
m_page_levenshtein=load_simple_json('de/wiki/m_page_levenshtein.json')
#n_page_levenshtein=load_simple_json('de/wiki/n_page_levenshtein.json')


page_f_validated=load_simple_json('de/wiki/f_page_validated.json')
page_f_ambigious=load_simple_json('de/wiki/f_page_ambigious.json')
page_f_other=load_simple_json('de/wiki/f_page_other.json')  
    
page_m_validated=load_simple_json('de/wiki/m_page_validated.json') 
page_m_ambigious=load_simple_json('de/wiki/m_page_ambigious.json')
page_m_other=load_simple_json('de/wiki/m_page_other.json')

page_n_validated=load_simple_json('de/wiki/n_page_validated.json')
page_n_ambigious=load_simple_json('de/wiki/n_page_ambigious.json')
page_n_other=load_simple_json('de/wiki/n_page_other.json') 

m_links_to_feminine=load_simple_json('de/wiki/m_links_to_feminine.json')
m_links_to_feminine_lev=load_simple_json('de/wiki/m_links_to_feminine_lev.json')
m_links_to_smth=load_simple_json('de/wiki/m_links_to_smth.json')
f_links_to_mascuilne=load_simple_json('de/wiki/f_links_to_mascuilne.json')
f_links_to_smth=load_simple_json('de/wiki/f_links_to_smth.json')
redirection_n=load_simple_json('de/wiki/n_redirects.json')
n_links_to_masculine=load_simple_json('de/wiki/n_links_to_masculine.json')
n_links_to_smth=load_simple_json('de/wiki/n_links_to_smth.json')

f_links_to_mascuilne_lev=load_simple_json('de/wiki/f_links_to_mascuilne_lev.json')
f_links_to_mascuilne_with_link=load_simple_json('de/wiki/f_links_to_mascuilne_with_link.json')
f_redirection=load_simple_json('de/wiki/f_redirection.json')
m_redirection=load_simple_json('de/wiki/m_redirection.json')


f_bias=load_simple_json('de/wiki/f_bias.json')
m_bias=load_simple_json('de/wiki/m_bias.json')
n_bias_male_labels=load_simple_json('de/wiki/n_bias_male_labels.json')
n_bias_n_labels=load_simple_json('de/wiki/n_bias_n_labels.json')
all_bias={words[i][1]: "female" for i in f_bias}
#all_bias2={i: "neutral" for i in n_bias_n_labels}
{all_bias.update({i: "neutral"}) for i in n_bias_n_labels}
{all_bias.update({i: "male"}) for i in m_bias}
{all_bias.update({i: "neutral"}) for i in n_bias_male_labels}
    

dict_m=[]
#df1=pd.DataFrame(columns=["Name","WikiName","Oppositgender","Bias"])
for m in words:
    if m in page_m_validated:
        dict_m.append({"Name":m,"WikiName":m,"Oppositgender":words[m][1],"Bias":all_bias[m],"Ambigious":0})
    elif words[m][1] in f_links_to_mascuilne:
        #print m
        dict_m.append({"Name":m,"WikiName":m,"Oppositgender":words[m][1],"Bias":all_bias[m],"Ambigious":1})
    elif m_page_levenshtein.has_key(m):
        #print m
        dict_m.append({"Name":m,"WikiName":m_page_levenshtein[m],"Oppositgender":words[m][1],
                       "Bias":all_bias[m],"Ambigious":0})
    elif f_links_to_mascuilne_with_link.has_key(words[m][1]) :
        #print m,f_links_to_mascuilne_with_link[words[m][1]][1]
        dict_m.append({"Name":m,"WikiName":f_links_to_mascuilne_with_link[words[m][1]][1],"Oppositgender":words[m][1],
                       "Bias":all_bias[m],"Ambigious":0})
    elif (m in page_m_ambigious)& (all_bias.has_key(m)):
        dict_m.append({"Name":m,"WikiName":m,"Oppositgender":words[m][1],"Bias":all_bias[m],"Ambigious":1})
    elif (m_links_to_smth.has_key(m))& (all_bias.has_key(m)):
        #print m
        dict_m.append({"Name":m,"WikiName":m_links_to_smth[m],"Oppositgender":words[m][1],"Bias":all_bias[m],"Ambigious":0})
    elif m_neutral_lev.has_key(m):
        print m,m_neutral_lev[m]
        dict_m.append({"Name":m,"WikiName":m_neutral_lev[m],"Oppositgender":words[m][1],"Bias":all_bias[m],"Ambigious":0})
        #pass
    if words[m][1] in page_f_validated:
        #print m
        dict_m.append({"Name":words[m][1],"WikiName":words[m][1],"Oppositgender":m,
                       "Bias":all_bias[m] if m in all_bias else all_bias[words[m][1]],"Ambigious":0})
    elif words[m][1] in page_f_ambigious:
        #print m
        dict_m.append({"Name":words[m][1],"WikiName":words[m][1],"Oppositgender":m,
                       "Bias":all_bias[m] if m in all_bias else all_bias[words[m][1]],"Ambigious":1})
for i in n_bias_n_labels:
    dict_m.append({"Name":i,"WikiName":i,"Bias":all_bias[i],"Ambigious":0})
df1=pd.DataFrame(dict_m)

#df1=df1.append(dict_m,ignore_index=True)   

df1[df1.Bias=="neutral"]#67=36neutral+16f_and_m+11n_redirects+4f

print len(df1)
df1.to_csv("de/wiki/all_wiki_pages.csv", sep='\t', encoding='utf-8')

import pandas as pd
import json
def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)
    
df=pd.read_csv('de/wiki/all_wiki_pages.csv',sep='\t', encoding='utf-8',index_col=0)
print len(df[(df.Bias=="male")] )
print len(df[(df.Bias=="male")& (df.Name!=df.WikiName)] )
print len(df[(df.Ambigious==1)] )
df[(df.Ambigious==1)]
