==> 8. Collect text and links and store all.ipynb <==
# Collect text and links from Wiki articles

Example of API queries:
#links to Wiki API all plane text and lead

https://de.wikipedia.org/w/api.php?action=query&prop=extracts&titles=Gleichstellungsbeauftragte&exintro=&explaintext=&redirects=&exsentences=10&formatversion=2

https://de.wikipedia.org/w/api.php?action=query&prop=extracts&titles=Gleichstellungsbeauftragte&exintro=&explaintext=&redirects=&formatversion=2

#from wikitools2 import wiki, api #new version of wikitools
from wikitools import wiki, api
import urllib, json
import re, urlparse
import pprint
import mwparserfromhell
from mwparserfromhell.nodes import Wikilink
import pandas as pd

def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)
    
def check_categories(page):
    categories_f=["Kategorie:Frau"]
    categories_m=["Kategorie:Mann"]
    #http://de.wikipedia.org/w/api.php?action=query&titles=Hebamme&prop=categories
    l="de" 
    site = wiki.Wiki("http://"+l+".wikipedia.org/w/api.php")
    params={'action':'query','titles':page,'prop':'categories','format':'json','continue':'','redirects':''}
    request = api.APIRequest(site, params)
    for result in request.queryGen():
        if result["query"]["pages"].values()[0].get("categories"):
            for i in result["query"]["pages"].values()[0]["categories"]:
                if i["title"] in categories_f:
                    return "Frau" 
                elif i["title"] in categories_m:
                    return "Mann"   
        else:
            return None   
    return None
def query_de_wiki(site,params):
    req = api.APIRequest(site, params)
    res=req.query()
    return res["query"]["pages"][0]["extract"]
    
def get_links_from_text(title):   
    language="de"
    site = wiki.Wiki("http://" + language + ".wikipedia.org/w/api.php")
    params = {'action':'parse', 'page':title, 'prop':'wikitext'}
    req = api.APIRequest(site, params) 
  
    alllinklist = []
    male_person=[]
    female_person=[]
    all_text=""
    lead_text=""
    
    #store all raw text
    params={'action':'query','titles':title,'prop':'extracts','explaintext':'','redirects':'','formatversion':'2'}
    all_text=query_de_wiki(site,params)
            
    #store lead raw text
    params['exintro']=''
    lead_text=query_de_wiki(site,params)
            
    if True:
        #try:
        #res= req.query(querycontinue=False) #False ?
        for res in req.queryGen():
            #wiki text with links and text as wiki markup language
            text = res['parse']['wikitext']['*']
            #a=mwparserfromhell.parse(text)
            #print a.strip_code()

            #parse all links
            filtered_code = mwparserfromhell.parse(text).filter_wikilinks()
            for node in filtered_code: 
                link = node.title
                if ":" not in link:
                    if link is not None:
                        strlink = link.encode("utf-8").decode("utf-8").replace(' ', '_')
                        # strlink=urllib.quote(strlink.encode("utf-8"))
                        if strlink not in alllinklist:
                            alllinklist.append(strlink)
                            #Kategorie:Frau, Kategorie:Mann
                            #check the category of the link
                            gender=check_categories(strlink)
                            if gender =="Mann":
                                male_person.append(strlink)
                            elif gender=="Frau":
                                female_person.append(strlink)
                                

        return alllinklist,male_person,female_person,all_text,lead_text  
            
Dump = {} 
data=pd.read_csv("de/wiki/all_wiki_pages.csv",sep='\t', encoding='utf-8',index_col=0)   
pages=data.WikiName.values  
#TODO: check whether file exist
with open('de/wiki/all_data_from pages.json', 'r') as in_f:
    Dump=json.load(in_f)
print "We already stored", len(Dump), "professions", "out of ",len(pages)
#pages=["Gleichstellungsbeauftragte","Lehrer","Arzt"]
for title in pages:
    if title not in Dump:
        alllink,m,f,all_text,lead_text = get_links_from_text( title)
        print len(alllink)
        Dump.update({title:{"all_link":alllink,"male_person":m,"female_person":f,"male_person_amount":len(m),
                                "female_person_amount":len(f),"text":all_text,"lead":lead_text}})             

with open('de/wiki/all_data_from pages.json', 'w') as out:
    json.dump(Dump, out, indent=4, sort_keys=True)  

# Store redirection bias ,gender of article and labour market in one file

import json
import pandas as pd

def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f,encoding="utf-8")
all_wiki_pages=pd.read_csv("de/wiki/all_wiki_pages.csv",sep='\t', encoding='utf-8',index_col=0)##885
about_pages=load_simple_json('de/wiki/all_data_from pages.json')
about_pages=pd.DataFrame.from_dict(about_pages,orient="index")#885

#==============add Bias
about_pages["Bias"]=about_pages.apply(lambda x:all_wiki_pages[all_wiki_pages.WikiName.isin([x.name])].Bias.values[0],axis =1)

#==============add profession gender
import copy
f_val=load_simple_json('de/wiki/f_page_validated.json')
f_amb=load_simple_json('de/wiki/f_page_ambigious.json')
n_pages=load_simple_json('de/wiki/n_page_validated.json')
n_pages2=all_wiki_pages[(all_wiki_pages.Bias=="neutral")&(all_wiki_pages.Name!=all_wiki_pages.WikiName)].WikiName.values#10 neutral names of professions
n_pages_all=copy.deepcopy(n_pages)
n_pages_all.extend(n_pages2)
f_val_all=copy.deepcopy(f_val)
f_val_all.extend(f_amb)

def identify_gender(f_val_all,n_pages_all,x):
    if x.name in f_val_all:
        return "female"
    elif x.name in n_pages_all:
        return "neutral"
    else:
        return "male"
about_pages["prof_gender"]=about_pages.apply(lambda x: identify_gender(f_val_all,n_pages_all,x),axis=1)
#==check data
#print len(n_pages2)+len(n_pages)#11+33
#print len(f_val_all)#18 (6bias+12)
#all_wiki_pages[all_wiki_pages.Bias.isin(["neutral"])]#f+n=73(6+67)
#about_pages[about_pages.index.isin(set(set(all_wiki_pages[all_wiki_pages.Bias.isin(["neutral"])].WikiName)-set(n_pages_all))-set(f_val_all))]
#print about_pages[(about_pages.Bias=="male")&(about_pages.prof_gender!="male")]
#print about_pages[(about_pages.Bias=="female")&(about_pages.prof_gender!="female")]


#==============add labour market data
all_together=pd.read_csv("de/all_labour.csv",sep='\t', encoding='utf-8',index_col=0) 
list_of_extra_names=list(all_wiki_pages[all_wiki_pages.WikiName!=all_wiki_pages.Name].Name.values)#48
list_of_extra_wikinames=list(all_wiki_pages[all_wiki_pages.WikiName!=all_wiki_pages.Name].WikiName.values)


#male
df_most_m=all_together[all_together.Frauen<50.0]
#wikipedia page names for male dominated professions
m_=list(all_wiki_pages[all_wiki_pages.Name.isin(df_most_m.name)].WikiName.values) #547
#female_names of male dominated professions
m_m=list(all_wiki_pages[all_wiki_pages.Oppositgender.isin(df_most_m.name)].WikiName.values) #6
print len(m_),len(m_m)
m_.extend(m_m)#553
print len(m_)#553

#another names of male prfessions 
m_most_m_other=df_most_m[df_most_m.name.isin(list_of_extra_names)].name#32
print len(m_most_m_other)#32
m_other=list(all_wiki_pages[all_wiki_pages.Name.isin(m_most_m_other)].WikiName.values)
m_.extend(m_other)#
print len(m_)#585
male_most_people=about_pages[about_pages.index.isin(m_)]#553
print len(male_most_people)


#female
#get where more than 50% of women involved
df_most_f=all_together[all_together.Frauen>=50.0]#311

print len(df_most_f)
f_=list(all_wiki_pages[all_wiki_pages.Name.isin(df_most_f.name)].WikiName.values) #306
#female names of female dominated professions
f_f=list(all_wiki_pages[all_wiki_pages.Oppositgender.isin(df_most_f.name)].WikiName.values) #12
print len(f_),len(f_f)
f_.extend(f_f)#318
print len(f_)
#other names f female professions
f_most_f_other=df_most_f[df_most_f.name.isin(list_of_extra_names)].name#7
f_other=list(all_wiki_pages[all_wiki_pages.Name.isin(f_most_f_other)].WikiName.values)
f_.extend(f_other)#
print len(set(f_)) #318
#309+535=844 over 885
female_most_people=about_pages[about_pages.index.isin(f_)]#318
print "",len(female_most_people)


print "No statistical data for ",len(list(set(set(about_pages.index)-set(m_))-set(f_)))," profesions:"
print list(set(set(about_pages.index)-set(m_))-set(f_))#14
#553+318=871 (871+14)=885

male_most_people["labor_market_domin"]=male_most_people.apply(lambda x: "male",axis=1)#"male"
female_most_people["labor_market_domin"]=female_most_people.apply(lambda x: "female",axis=1)#"female"
#all data
df_all_labour=pd.concat([male_most_people,female_most_people])

def do_smth(x):
    if x.name in df_all_labour.index.values:
        return df_all_labour[df_all_labour.index==x.name].labor_market_domin.values[0]
    else:
        return None
def return_woman_ratio_in_labour(x):
    #TODO
    if x.name in df_all_labour.index.values:
        return smth
    else:
        return None    

about_pages["labor_market_domin"]=about_pages.apply(lambda x: do_smth(x),axis=1)
about_pages["profession"]=about_pages.apply(lambda x: x.name,axis=1)
#about_pages["labor_market_woman"]=about_pages.apply(lambda x: return_woman_ratio_in_labour(x),axis=1)
#!!!!Store it in file==================================!!!!!!!!!!!!!!=====================
about_pages.to_csv('de/all_data_bias_labour_gender.csv', sep='\t', encoding='utf-8')

about_pages

print "No labour market data for these professions (they are too general):"
about_pages[about_pages.labor_market_domin.isnull()]
