==> 9. Extract pictures.ipynb <==
#import wikipedia
from wikitools import wiki, api
#import MySQLdb
import re
import csv
import pandas as pd

import json
def load_simple_json(filename):
    with open(filename, 'r') as f:
        return json.load(f)

class professions:
    rows_list=[]
    job_list=[]


    def get_image_link(self,language=1):
        wiki_image_list=[]
        
        df = pd.DataFrame(columns=[ 'language_label', 'profession','wikipedia_link', 'image_link','image_name'])
        n=0
        bad_pictures=["Datei:Disambig-dark.svg","Datei:Commons-logo.svg","Datei:Wikisource-logo.svg",
                      "Datei:Wikiquote-logo.svg","Datei:Wikiversity-logo.svg","Datei:Wikibooks-logo.svg",
                      "Datei:Flag of Germany.svg","Datei:Icon-Rechtshinweis-blau2-Asio.svg"]
        combined_bad = '\t'.join(bad_pictures)
        for i in self.rows_list:
            if False:#n>=10:
                break
            n+=1
            all_links=[]

            prof = i
            try:
                page_title = i
            except:
                continue
            print page_title
    
            params = {'action':'query',
                'format':'json',
                'titles':page_title,
                'prop':'langlinks',
                'lllimit':'100',
                'llprop':'url',
                "iwurl":'',
                'continue':''
            }
            try:
                if language==1:
                    filename = 'profession_images_all.csv'
                    all_links.append(['de',page_title,prof,'http://de.wikipedia.org/wiki/'+page_title])
                else:
                    filename = 'profession_images_all_languages.csv'
                    #print filename
                    #---------------------------------- wikitools library ----------------------------------------
                    site = wiki.Wiki('http://de.wikipedia.org/w/api.php')
                    req = api.APIRequest(site, params)
                    res = req.query()
                    # pprint.pprint(res)
                    #--------------------------------------------------------------------------------------
                    #print res
                    all_links.append(['de',page_title,prof,'http://de.wikipedia.org/wiki/'+page_title])

                    res1 = res['query']['pages'].values()[0]
                    if res1.get("langlinks"):
                        lang_list=res1['langlinks']  
                        for lang in lang_list:
                            lang_id = lang['lang']
                            lang_url = lang['url']
                            wiki_name = lang['*']

                            all_links.append([lang_id,wiki_name,prof,lang_url])
                    
                for data in all_links:
                    lang_id = data[0]
                    try :

                        #print data[1]
                        params2 = {'action':'query',
                            'format':'json',
                            'titles':data[1],
                            'generator':'images',
                            'gimlimit':10,
                            'prop':'imageinfo',
                            'iiprop':'url|size',
                                   'iiurlwidth':500,
                            'continue':''
                        }

                        #---------------------------------- wikitools library ----------------------------------------
                        site2 = wiki.Wiki('http://%s.wikipedia.org/w/api.php' %lang_id)
                        req = api.APIRequest(site2, params2)
                        res = req.query()
                        #--------------------------------------------------------------------------------------
                        for item in  res['query']['pages'].values():
                            #if 'Wiktionary' or 'Ambox' not in item['imageinfo'][0]['url'].split('/')[7]:
                                #print item['imageinfo'][0]['url']
                            if (item['imageinfo'][0]['width']>100):#&(item["title"] not in combined_bad):
                                if ".tif" in item["title"]:
                                    img_link=item['imageinfo'][0]['thumburl']
                                else:
                                    img_link=item['imageinfo'][0]['url']
                                wiki_dict={ 'language_label':data[0], 'profession':data[2],'wikipedia_link':data[3],
                                    'image_link':img_link,'image_name':item["title"].rsplit(":",1)[-1]}
                                df = df.append(wiki_dict, ignore_index=True)
                        if df["profession"].iloc[-1]!=data[2]:
                            wiki_dict={ 'language_label':data[0], 'profession':data[2],'wikipedia_link':data[3],
                                       'image_link':'-','image_name':None}
                            df = df.append(wiki_dict, ignore_index=True)
                            
                    except:
                        wiki_dict={ 'language_label':data[0], 'profession':data[2],'wikipedia_link':data[3],
                                   'image_link':'-','image_name':None}
                        df = df.append(wiki_dict, ignore_index=True)
                        continue

            except:
                print "error on row : " ,i
                continue
        df.to_csv(filename, sep='\t', encoding='utf-8')

    def search(self):
        counter =0
        for i in range(len(self.rows_list)):
            counter +=1
            print counter
            prof = self.rows_list[i][0][:-1]
            result = wikipedia.search(prof)
            print prof, " : ",result
            for item in result:
                #print prof
                try:
                    if prof in item:
                        print item
                        self.job_list.append(item)
                except:
                    continue
        print len(self.job_list)
if __name__ == '__main__':
    p = professions()
    data=pd.read_csv("de/wiki/all_wiki_pages.csv",sep='\t', encoding='utf-8',index_col=0)
    p.rows_list=data.WikiName.values
    #p.rows_list=load_simple_json("de/wiki/m_page_validated.json")
    p.get_image_link()
    #p.get_image_link(language=99)
    
    
    


# Old statistics. Now new one in the file: 10. Find matching images from Claudia's project.ipynb

[open notebook 10. ...](10. Find matching images from Claudia's project.ipynb)

import pandas as pd

df_images=pd.read_csv("profession_images_all.csv",sep='\t', encoding='utf-8',index_col=0)
print "No images for ",len(df_images[df_images.image_link=="-"]),"professions"
df_with_image=df_images[df_images.image_link!="-"]
print "We have approx",len(df_with_image)/float(len(pd.unique(df_with_image.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df_with_image.profession)),"professions with image."

df_with_image_without_svg=df_with_image[~df_with_image.image_link.str.contains(".svg")]
print "Images without .svg:",len(df_with_image_without_svg)
print "We have approx",len(df_with_image_without_svg)/float(len(pd.unique(df_with_image_without_svg.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df_with_image_without_svg.profession)),"professions with image."

video=df_with_image[df_with_image.image_link.str.contains(".ogg|.ogv")] 
print "There are ",len(video)," videos."
video

df_with_image_without_svg_and_video=df_with_image_without_svg[~df_with_image_without_svg.image_link.str.contains(".ogg|.ogv")]
print "Images without .svg and video:",len(df_with_image_without_svg_and_video)
print "We have approx",len(df_with_image_without_svg_and_video)/float(len(pd.unique(df_with_image_without_svg_and_video.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df_with_image_without_svg_and_video.profession)),"professions with image."
