==> 10. Find matching images from Claudia's project.ipynb <==
# Find images for wich we should start CrowdFlower Project

## We also exclude video and svg files from image list

#add manually "Hannah Aitchison.jpg" according to 15th of February data
#=====
import pandas as pd
df_images=pd.read_csv("profession_images_all.csv",sep='\t', encoding='utf-8',index_col=0)
df_images_previous=pd.read_csv("profession_images.csv",sep='\t', encoding='utf-8',index_col=0)
if "Hannah Aitchison.jpg" not in df_images.image_name.values:
    df_images=df_images.append(df_images_previous[df_images_previous.image_name=="Hannah Aitchison.jpg"]).reset_index(drop=True)
    df_images.to_csv("profession_images_all.csv", sep='\t', encoding='utf-8')
#========


import pandas as pd
import urllib2


df_images=pd.read_csv("profession_images_all.csv",sep='\t', encoding='utf-8',index_col=0)

print "No images for ",len(df_images[df_images.image_link=="-"]),"professions"
df_with_image=df_images[df_images.image_link!="-"]
print "We have approx",len(df_with_image)/float(len(pd.unique(df_with_image.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df_with_image.profession)),"professions with image."
print "Now we will=====>We exclude .svg images abd video."
df_with_image_without_svg=df_with_image[~df_with_image.image_link.str.contains(".svg")]
video=df_with_image[df_with_image.image_link.str.contains(".ogg|.ogv")] 
print "There are ",len(video)," videos."
df=df_with_image_without_svg[~df_with_image_without_svg.image_link.str.contains(".ogg|.ogv")]
print "Such that, no images for ",len(pd.unique(df_images.profession))-len(pd.unique(df.profession)),"professions"
print "Images without .svg and video:",len(df)
print "We have approx",len(df)/float(len(pd.unique(df.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df.profession)),"professions with image."


df_results=pd.read_csv("crowdflower-results-v3.csv",sep=';', encoding='utf-8',index_col=0)
df_results=df_results[~df_results.image_url.isnull()]
df_results=df_results[df_results.no_of_persons!="Image is not shown"]
url_of_results=df_results.image_url.values

#check whether same_link and url_of_results are the same
import urllib2

k=0
same_links_k=[]
old_link_name_new={}
sl=set(df.image_name.str.replace(" ","_"))
p=0
for i in url_of_results:
    for j in sl:
        if True:
            j_quote=urllib2.quote(j.encode('utf-8'))
            if j_quote in i:  
                k+=1
                same_links_k.append(urllib2.unquote(j.replace("_"," ").encode('utf-8')).decode('utf-8'))
                old_link_name_new[i]=urllib2.unquote(j.replace("_"," ").encode('utf-8')).decode('utf-8')

new_links=df[~df.image_name.isin(same_links_k)]
same_links=df[df.image_name.isin(same_links_k)]
print "We have ",len(same_links), "same links for ",len(pd.unique(same_links.profession))," unique professions."

print "We have ",len(new_links),"new links for ",len(pd.unique(new_links.profession))," professions."
df_unique_links2=new_links[~new_links.duplicated('image_link')].image_link.to_frame()
df_unique_links2.columns=["image_url"]
print "We have ",len(df_unique_links2),"new links for "

#save old links from Caludia_project, we will use the results in our research
import json
with open('de/old_link_name_new.json', 'w') as fp:
    json.dump(old_link_name_new, fp,encoding='utf-8')

# Save all image links

df_unique_links2.to_excel("Crowdflower_image_links_all.xls",header=["image_url"], index=False,encoding='utf-8')

# Add difference between "profession_images.csv" and "profession_images_all.csv" 
#Save it to file "Crowdflower_image_links_more.xls"

#find difference mit previously created list and new one.


df_unique_links_previous=pd.read_excel("Crowdflower_image_links.xls",index_col=None, encoding='utf-8')
difference=set(df_unique_links_previous.image_url)-set(df_unique_links2.image_url)
print set(df_unique_links_previous.image_url)-set(df_unique_links2.image_url)
print len(set(df_unique_links2.image_url)-set(df_unique_links_previous.image_url))


#compare "Crowdflower_image_links.xls" and new list 
df_unique_links2=df_unique_links2.reset_index(drop=True)
df_unique_links_previous=df_unique_links_previous.reset_index(drop=True)
df =pd.concat([df_unique_links2, df_unique_links_previous],ignore_index=True)
print len(df_unique_links2)+len(df_unique_links_previous),len(df)

df_gpby = df.groupby(list(df.columns))
idx = [x[0] for x in df_gpby.groups.values() if len(x) == 1]
diff=df.reindex(idx)

diff=diff[~diff.image_url.isin(list(difference))]


print "Amount of new images",len(diff)
diff.to_excel("Crowdflower_image_links_more.xls",header=["image_url"], index=False,encoding='utf-8')

# print difference

set(df_unique_links2.image_url)-set(df_unique_links_previous.image_url)

# End is here! Other stuff is old ==> Do not use!

import pandas as pd

df_images=pd.read_csv("profession_images.csv",sep='\t', encoding='utf-8',index_col=0)
print "No images for ",len(df_images[df_images.image_link=="-"]),"professions"
df_with_image=df_images[df_images.image_link!="-"]
print "We have approx",len(df_with_image)/float(len(pd.unique(df_with_image.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df_with_image.profession)),"professions with image."

print "We exclude .svg images abd video."
df_with_image_without_svg=df_with_image[~df_with_image.image_link.str.contains(".svg")]
video=df_with_image[df_with_image.image_link.str.contains(".ogg|.ogv")] 
print "There are ",len(video)," videos."
df=df_with_image_without_svg[~df_with_image_without_svg.image_link.str.contains(".ogg|.ogv")]
print "Such that, no images for ",len(pd.unique(df_images.profession))-len(pd.unique(df.profession)),"professions"
print "Images without .svg and video:",len(df)
print "We have approx",len(df)/float(len(pd.unique(df.profession))),\
"pictures per profession with image. "+"We have",len(pd.unique(df.profession)),"professions with image."

df_images_another_project=pd.read_csv("profession_images (2).txt",sep='\t', encoding='utf-8',index_col=0)
de_images_another_project=df_images_another_project[df_images_another_project.language_label=="de"]
de_images_another_project=de_images_another_project[de_images_another_project.image_link!="-"]
#print "Number of images from de wiki",len(de_images_another_project),"for",len(pd.unique(de_images_another_project.profession))," professions."
new_links=df[~df.image_link.isin(df_images_another_project.image_link.values)]
print "We have ",len(new_links),"new links for ",len(pd.unique(new_links.profession))," professions."
same_links=df[df.image_link.isin(df_images_another_project.image_link.values)]
print "We have ",len(same_links), "same links for ",len(pd.unique(same_links.profession))," unique professions."
same_links[same_links.image_link.str.contains(".tif")]

#check the result file
df_results=pd.read_csv("crowdflower-results-v3.csv",sep=';', encoding='utf-8',index_col=0)
df_results=df_results[~df_results.image_url.isnull()]
df_results=df_results[df_results.no_of_persons!="Image is not shown"]
url_of_results=df_results.image_url.values


#check whether same_link and url_of_results are the same
import urllib2

k=0
sl=df.image_link.str.rpartition("/")[2].values#.isin(url_of_results)
for i in url_of_results:
    for j in sl:
        if j in i:
            k+=1
print k,len(same_links)
df_unique_links=new_links[~new_links.duplicated('image_link')].image_link.to_frame()
print "We have",len(df_unique_links),"unique links."
df_unique_links.to_excel("Crowdflower_image_links.xls",header=["image_url"], index=False,encoding='utf-8')

#old!!! do not use it 


d_set=set(same_links.image_name)-set(same_links_k)
new_new=same_links[same_links.image_name.isin(d_set)]
new_newunique_links=new_new[~new_new.duplicated('image_link')].image_link.to_frame()
print new_newunique_links[new_newunique_links.image_link.isin(df_unique_links_previous.image_url.values)]
print new_newunique_links[new_newunique_links.image_link.isin(df_unique_links.image_url.values)]
new_newunique_links.to_excel("Crowdflower_image_links2.xls",header=["image_url"], index=False,encoding='utf-8')
